Added support for JSON output
Working on integrating digCUST with blueplanet so added the option of
displaying output to json so that blueplanet can communicate to digCUST
on a REST interface.

Also changed some of the gsub patterns in the awk files that just match
for /"/, changed all instances of that to "\"". This was done because
in most IDEs this breaks the syntax highlight of strings so doing
gsub on a string pattern as opposed to a regex object fixes the
syntax highlighting.