Updated digCUST shell script
Updated digCUST shell script so that it gives priority formatting to
JSON just below raw so that it if it used it will ignore any other
output type like csv and web but will be ignored if raw is used.
It will also turn off ascii formatting if it used.