Updated movedigcust script
Had a statement that was setting the digCUST directory to a static
directory which was the same on 2 hosts but became an issue on the 3rd
that got added recently in the move operation. Fixed the variable to
be dynamic and will be relative to where it was called.d.