Changed expect prompts in checksshkeys.tcl
The checksshkeys.tcl script was always checking for the pattern
{ltrkarkvm066:DigCust \$} for when it started up the bash shell and
when it logged out of the devices it logged into as the patter for the
prompt. Thing is there are 2 other repos for this project like this one
named digCUST and one called digCust. Changed the pattern to match all.